chore(model): add `model_spec` and `model_instance_spec` in `ModelDefinition` (#67)
Because

A model and a model instance from the same model definition may require different fields. Use the model definition `GitHub` as an example, to create a model, a repository is required. But for each imported model instance, an extra `tag` field is needed. Therefore, for each supported model definition, we provide specifications for both model and model instance.

This commit

- remove `public` and `custom` fields from `ModelDefinition` message
- remove `Spec` message and add `model_spec` and `model_instance_spec` in `ModelDefinition` message