feat(vdp): remove openapi_specification and add data_specification (#282)
Because

- Currently, we have an `openapi_specification` field in the pipeline
and connector/operator definition. The original purpose was that we
could generate an OpenAPI schema and have some UI to let the user debug
on an individual component. However, it turns out that we don't really
need that OpenAPI schema, and it makes our response very large and hard
to read.

This commit

- Simplifies `openapi_specification` into `data_specification` by
preserving only the input and output JSON schema and removing the nested
structure of the OpenAPI spec.
- Adds `data_specifications` in iterator component.

---------

Co-authored-by: droplet-bot <droplet-bot@users.noreply.github.com>