feat(artifact): add tag create / list endpoints (#286)
Because

- We want to list the versions of a model.
- Registry has a
[tags](https://distribution.github.io/distribution/#listing-image-tags)
endpoint but it doesn't expose any timestamp.

![CleanShot 2024-03-19 at 20 24
14](https://github.com/instill-ai/protobufs/assets/3977183/38f408d1-0096-4f39-b3fe-92991104fa66)

This commit

- Adds and documents a private API for the repository tags in artifact.
- Only `api-gateway` (when it detects the end of the image push
operation) should communicate the creation of a tag.
- The tags won't be exposed directly to the client. `model-backend` will
request this information to expose the model versions in the same family
as the rest of the model endpoints.

## 🗒️ Notes

As an initial addition to the Artifact domain, I made some naming
choices. Please review and provide feedback about them.

## ⏭️ Next steps

- `artifact-backend` will implement a first version of this endpoint
that will fetch the tag list from the `registry` instance. This endpoint
only returns the tag list so the digest and update time will be empty.
- In order to record the update time, `api-gateway` will call the
creation endpoint when we intercept a `PUT manifest` request. We'll
store the digest and update time in the `artifact` database and we'll
aggregate that info to the `registry` tag list (the registry should
remain the source of truth).

---------

Co-authored-by: droplet-bot <droplet-bot@users.noreply.github.com>