feat(vdp): refactor pipeline recipe and optimize the validation endpoint (#341)
Because

- Originally, we used protobuf messages to fully describe the recipe,
which is not suitable for our new recipe design. The protobuf `oneof`
would introduce an additional nested structure in JSON when using
grpc-gateway. This adds extra effort for users when they want to
directly write a JSON recipe. We have decided to remove the messages,
and the recipe format will now be validated by JSON-schema in the
pipeline-backend instead. And in the near future, we'll also change to
use YAML recipes, which will be a string and cannot be validated by
protobufs anyway.
- We need to provide an endpoint for recipe validation that can return
all errors in the recipe.

This commit

- Removes all recipe-related messages and use `structpb` for the recipe
instead.
- Updates the response format of `validate` endpoint.

Note:
When using grpc-gateway, the corresponding JSON format for the oneof
fields will be divided into different nested objects. This is not the
same as the `oneOf` in JSON schema that we want, which don't have the
nested structure.

protobuf
```
message XXX {
oneof component {
    // ConnectorComponent
    ConnectorComponent connector_component = 13;
    // OperatorComponent
    OperatorComponent operator_component = 14;
    // IteratorComponent
    IteratorComponent iterator_component = 15;
  }
}

```

JSON by grpc-gateway
```
{
  "connector_component": {},
  "operator_component": {},
  "iterator_component": {},
}
```

---------

Co-authored-by: droplet-bot <droplet-bot@users.noreply.github.com>