feat(mgmt): unify credit and pipeline metric endpoints (#387)
Because

- A new dashboard page is going to be implemented, where credit, model
and pipeline metrics are displayed.

This commit

- Unifies the shape of the credit and pipeline endpoints in order to
reduce the complexity on the client and server.
- Removes unused fields / endpoints.
- The source of truth for the trigger list isn't the metrics endpoint
anymore, the logging domain is. The dashboard will redirect to the
trigger list powered by the log tables.
- We don't need the success / failure count _per pipeline_ anymore.
Instead of a `*TableRecords` endpoint, a new `Count` endpoint is added.
- The `*ChartRecords` endpoints are unified. All the params that are
needed in the query (all those that were required or had a default
value) are explicit. Though the dashboard won't have a filter initially,
we might filter by other fields in the future through an AIP filter
param.

---------

Co-authored-by: droplet-bot <droplet-bot@users.noreply.github.com>