chore(agent): add object UID for chat (#586)
Because

- Currently, the chat API accepts a file UID as input, but this requires
the frontend to bind the file to the catalog first. We’d like to
simplify this flow so that the frontend doesn’t need to manage the
catalog. Therefore, we need to allow users to send the file using an
object UID, which doesn’t need to be bound to the catalog.

This commit

- Uses the object UID as input.